# frontend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AeroLicense API",
    "description": "Kurgusal bir Sivil Havacılık Otoritesi için lisans ve uçuş kaydı prototipi.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/training-records": {
      "post": {
        "tags": [
          "TrainingRecords"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingRecordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingRecordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingRecordRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingRecordDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "TrainingRecords"
        ],
        "parameters": [
          {
            "name": "ApplicantId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "PassedOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingRecordDtoPagedResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTrainingRecordRequest": {
        "required": [
          "applicantNationalId",
          "completedAtUtc",
          "examScore",
          "licenseType"
        ],
        "type": "object",
        "properties": {
          "applicantNationalId": {
            "type": "string"
          },
          "licenseType": {
            "$ref": "#/components/schemas/LicenseType"
          },
          "completedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "examScore": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "LicenseType": {
        "enum": [
          "PPL",
          "CPL",
          "ATPL"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "TrainingRecordDto": {
        "required": [
          "applicantId",
          "applicantName",
          "completedAtUtc",
          "examScore",
          "id",
          "isPassed",
          "licenseType",
          "trainingOrgName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "applicantId": {
            "type": "string",
            "format": "uuid"
          },
          "applicantName": {
            "type": "string"
          },
          "applicantNationalIdMasked": {
            "type": "string",
            "nullable": true
          },
          "trainingOrgName": {
            "type": "string"
          },
          "licenseType": {
            "$ref": "#/components/schemas/LicenseType"
          },
          "completedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "examScore": {
            "type": "integer",
            "format": "int32"
          },
          "isPassed": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "TrainingRecordDtoPagedResult": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainingRecordDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "required": [
          "errors"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": {}
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "POST /api/v1/auth/login'den alınan access token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}